requirements.txt
Django>=5.2,<6
django-silk==5.3.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
whitenoise>=6.5.0
gunicorn>=20.1.0
django-storages[boto3]
boto3>=1.26
python-dotenv==1.1.0

# дополнительные зависимости
pillow==11.2.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
gprof2dot==2025.4.14
inflection==0.5.1
pycodestyle==2.13.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1

# если подключаете celery (опционально)
django-celery-beat
django-celery-results
